pyproject.toml
[tool.poetry]
name = "fastapi_templates"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
alembic = "^1.12.0"
sqlmodel = "^0.0.9"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"
cryptography = "^41.0.4"
httpx = "^0.25.0"
factory-boy = "^3.3.0"
stringcase = "^1.2.0"
uvicorn = "^0.23.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
